Azure - Managed HSM KeyVault Fixes (#43)
* Allow specifying Azure Resource

When using a Managed HSM-backed KeyVault, Azure places this type of
resource under a different URI (managedhsm.azure.net) than the regular
Key Vault (vault.azure.net). This requires a different bearer token and
thus requires we identify the resource type via config. Allow specifying
AZ_RESOURCE_TYPE or the 'resource' key in the config to identify this.

Co-authored-by: Rachel Culpepper <84159930+rculpepper@users.noreply.github.com>
Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>

* Workaround Azure Managed HSMs API bug

Azure Managed HSMs currently modify the request body of the /wrapkey
(or potentially the response body of the /unwrapkey) API endpoint to
replace URLSafe Base64 padding characters with NULL bytes. When
unwrapping the default AES 256-bit symmetric key, we end up with a
spare NULL byte as it is not truncated by the API.

This results in the following error in Vault:

> failed to unseal barrier: failed to check for keyring: failed to create cipher: crypto/aes: invalid key size 33

If we detect the key has the invalid length, and the key ends with a
NULL byte, trim it before returning the value. Note that AES 128-bit
keys get two padding characters (and thus two NULL bytes) and AES
192-bit keys get no padding characters and thus no NULL bytes.

Co-authored-by: Rachel Culpepper <84159930+rculpepper@users.noreply.github.com>
Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>

Co-authored-by: Rachel Culpepper <84159930+rculpepper@users.noreply.github.com>