feature (extras/kms): add version to the entire kms collection
Add support for versioning the entire "collection" of keys represented
by the kms.

Any time one of the keys in the collection changes, the
version will be incremented.

Adding this capability will make it easy to invalidate the kms cache.
This is especially important when the cache is shared (distributed)
across multiple processes.

This change requires incrementing the migrations version.