feat(extras/crypto): Add Sha256SumReader (#165)
This adds an io.Reader implementation that is analogous to the
Sha256SumWriter. It can be used to read from an io.Reader and product a
sum. It also implements io.Closer so that if the provided io.Reader
implements io.Closer it can be called, if it does not, a call to Close
on the Sha256SumReader will be a noop.