Remove extras unused by OpenBao (#12)
These subpackages are unused by OpenBao. KMS is a package built on top
of go-kms-wrapping which uses a database to store encryption keys: this
is unnecessary in our case as we explicitly want all keys to be backed
by the underlying wrapper (which, admittedly in the case of cloud KMSes,
incurs some cost that extras/kms/ could avoid). Crypto adds, along other
things, a HMAC-SHA-256 implementation: if we find this useful, we could
add this directly to wrappers/ as an optional type (just like signing.go
is). Lastly, StructWrapping adds a way to encrypt arbitrary interfaces,
through a custom marshaling format built on protobuf.

extras/multi remains: it is unclear if it might potentially be useful
for multi-unseal in OpenBao in the future and aead/ had tests using it.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>